PML-323/tests: moved tests of deprecated methods to test_kernels_deprecation
