+++ tests/algorithms/test_kernels.py
@@ -342,27 +342,6 @@ class TestFidelityKernel:
         assert kernel.is_trainable
         assert "_quantum_layer.theta" in dict(kernel.named_parameters())
 
-    def test_kernel_rejects_no_bunching(self):
-        with pytest.warns(DeprecationWarning):
-            with pytest.raises(ValueError) as exc_info:
-                FidelityKernel(
-                    feature_map=self.feature_map,
-                    input_state=[2, 0],
-                    no_bunching=True,
-                )
-        assert "no_bunching" in str(exc_info.value)
-
-        with pytest.warns(DeprecationWarning):
-            with pytest.raises(ValueError) as exc_info:
-                FidelityKernel.simple(input_size=2, no_bunching=True)
-        assert "no_bunching" in str(exc_info.value)
-
-        builder = KernelCircuitBuilder().input_size(2).n_modes(4)
-        with pytest.warns(DeprecationWarning):
-            with pytest.raises(ValueError) as exc_info:
-                builder.build_fidelity_kernel(no_bunching=True)
-        assert "no_bunching" in str(exc_info.value)
-
     def test_kernel_scalar_computation(self):
         x1 = torch.tensor([0.5, 1.0])
         x2 = torch.tensor([1.0, 0.5])
@@ -446,68 +425,6 @@ class TestFidelityKernel:
                 computation_space=ComputationSpace.FOCK,
             )
 
-    def test_kernel_warns_and_uses_feature_map_encoder(self):
-        circuit = pcvl.Circuit(3)
-        for mode in range(3):
-            circuit.add(mode, pcvl.PS(pcvl.P(f"x{mode}")))
-
-        def encoder(x):
-            return torch.stack([x[0], x[1], x[0] + x[1]])
-
-        feature_map = FeatureMap(
-            circuit=circuit,
-            input_size=2,
-            input_parameters="x",
-            encoder=encoder,
-        )
-
-        with pytest.warns(
-            DeprecationWarning,
-            match="FeatureMap.encoder support inside FidelityKernel is deprecated",
-        ) as warning_record:
-            kernel = FidelityKernel(
-                feature_map=feature_map,
-                input_state=[1, 0, 0],
-                computation_space=ComputationSpace.FOCK,
-            )
-        warning_message = str(warning_record[0].message)
-        assert "CircuitBuilder.add_angle_encoding" in warning_message
-        assert "pre-encoding the data" in warning_message
-        assert "input_size equal to the encoded circuit-parameter count" in warning_message
-
-        encoded = kernel._quantum_layer._encode_single(torch.tensor([0.2, 0.3]))
-        expected = torch.tensor([0.2, 0.3, 0.5], dtype=encoded.dtype)
-        assert torch.allclose(encoded, expected)
-
-    def test_kernel_warns_and_uses_direct_circuit_subset_expansion(self):
-        circuit = pcvl.Circuit(3)
-        for mode in range(3):
-            circuit.add(mode, pcvl.PS(pcvl.P(f"x{mode}")))
-
-        feature_map = FeatureMap(
-            circuit=circuit,
-            input_size=2,
-            input_parameters="x",
-        )
-
-        with pytest.warns(
-            DeprecationWarning,
-            match="input_size differs from the circuit input parameter count",
-        ) as warning_record:
-            kernel = FidelityKernel(
-                feature_map=feature_map,
-                input_state=[1, 0, 0],
-                computation_space=ComputationSpace.FOCK,
-            )
-        warning_message = str(warning_record[0].message)
-        assert "CircuitBuilder.add_angle_encoding" in warning_message
-        assert "pre-encoding the data" in warning_message
-        assert "input_size equal to the encoded circuit-parameter count" in warning_message
-
-        encoded = kernel._quantum_layer._encode_single(torch.tensor([0.2, 0.3]))
-        expected = torch.tensor([0.2, 0.3, 0.5], dtype=encoded.dtype)
-        assert torch.allclose(encoded, expected)
-
     def test_kernel_uses_builder_subset_encoding_without_deprecation(self):
         builder = CircuitBuilder(n_modes=3)
         builder.add_angle_encoding(
@@ -551,18 +468,6 @@ class TestFidelityKernel:
             input_state=[1, 0, 0],
             computation_space=ComputationSpace.FOCK,
         )
-
-        encoded = kernel._quantum_layer._encode_single(torch.tensor([0.2, 0.4]))
-        expected = torch.tensor([0.1, 0.2], dtype=encoded.dtype)
-        assert torch.allclose(encoded, expected)
-
-    def test_simple_kernel_backend_preserves_angle_encoding_scale(self):
-        with pytest.warns(DeprecationWarning, match="n_modes"):
-            kernel = FidelityKernel.simple(
-                input_size=2,
-                n_modes=4,
-                angle_encoding_scale=0.5,
-            )
 
         encoded = kernel._quantum_layer._encode_single(torch.tensor([0.2, 0.4]))
         expected = torch.tensor([0.1, 0.2], dtype=encoded.dtype)
@@ -833,31 +738,6 @@ class TestFeatureMapFactoryMethods:
         assert feature_map.input_size == 3
         assert feature_map.circuit.m == 4  # input_size + 1
 
-    def test_simple_n_modes_override_emits_deprecation_warning(self):
-        """Passing n_modes to FeatureMap.simple must emit a DeprecationWarning."""
-        with pytest.warns(DeprecationWarning, match="n_modes"):
-            feature_map = FeatureMap.simple(input_size=2, n_modes=6)
-        assert feature_map.circuit.m == 6
-
-    def test_simple_factory_raises_when_input_exceeds_modes(self):
-        with pytest.warns(DeprecationWarning):
-            with pytest.raises(
-                ValueError, match="You cannot encore more features than mode with Builder"
-            ):
-                FeatureMap.simple(input_size=5, n_modes=4)
-
-    def test_simple_factory_raises_when_input_or_modes_exceeds_20(self):
-        with pytest.raises(ValueError):
-            FeatureMap.simple(input_size=21)
-        with pytest.warns(DeprecationWarning):
-            with pytest.raises(ValueError):
-                FeatureMap.simple(input_size=21, n_modes=21)
-
-    def test_simple_num_photons_modes_and_input_state(self):
-        for i in range(1, 15):
-            kernel = FeatureMap.simple(input_size=i)
-            assert kernel.circuit.m == i + 1
-
     def test_simple_trainable(self):
         for i in range(1, 20):
             kernel = FeatureMap.simple(input_size=i)
@@ -912,180 +792,11 @@ class TestFidelityKernelFactoryMethods:
         assert kernel.shots == 1000
         assert kernel.sampling_method == "multinomial"
 
-    def test_simple_factory_method(self):
-        """Test the simple FidelityKernel factory method."""
-        with pytest.warns(DeprecationWarning, match="n_modes"):
-            kernel = FidelityKernel.simple(input_size=2, n_modes=4)
-
-        assert kernel.input_size == 2
-        assert kernel.feature_map.circuit.m == 4
-        assert len(kernel.input_state) == 4
-        assert sum(kernel.input_state) == 2
-        assert kernel.input_state == [1, 0, 1, 0]
-
-    def test_simple_factory_default_photons(self):
-        """Test simple factory with default n_modes (should equal input_size + 1)."""
-        kernel = FidelityKernel.simple(input_size=3)
-
-        assert kernel.input_size == 3
-        assert kernel.feature_map.circuit.m == 4  # input_size + 1
-        assert sum(kernel.input_state) == 2
-        assert kernel.input_state == [1, 0, 1, 0]
-
-    def test_simple_num_photons_modes_and_input_state(self):
-        for i in range(1, 15):
-            kernel = FidelityKernel.simple(input_size=i)
-            assert kernel.feature_map.circuit.m == i + 1
-            assert np.sum(kernel.input_state) == int(np.ceil((i + 1) / 2))
-            assert len(kernel.input_state) == i + 1
-
-            input_state = [0] * (i + 1)
-            for j in range(len(input_state)):
-                if j % 2 == 0:
-                    input_state[j] = 1
-            assert kernel.input_state == input_state
-        for i in range(1, 15):
-            with pytest.warns(DeprecationWarning, match="n_modes"):
-                kernel = FidelityKernel.simple(input_size=1, n_modes=i + 1)
-            assert kernel.feature_map.circuit.m == i + 1
-            assert np.sum(kernel.input_state) == int(np.ceil((i + 1) / 2))
-            assert len(kernel.input_state) == i + 1
-
-            input_state = [0] * (i + 1)
-            for j in range(len(input_state)):
-                if j % 2 == 0:
-                    input_state[j] = 1
-            assert kernel.input_state == input_state
-
-    def test_simple_parameters(self):
-        for i in range(1, 15):
-            kernel = FidelityKernel.simple(input_size=i)
-            params = list(kernel.parameters())
-            named_params = [i[0] for i in kernel.named_parameters()]
-            assert params[0].numel() == i * (i + 1)
-            assert params[1].numel() == i * (i + 1)
-            assert len(params) == 2
-            assert kernel.feature_map.is_trainable
-            assert "_quantum_layer.LI_simple" in named_params
-            assert "_quantum_layer.RI_simple" in named_params
 
 
-class TestKernelCircuitBuilder:
-    """Test the KernelCircuitBuilder fluent interface."""
 
-    def test_builder_basic_usage(self):
-        """Test basic KernelCircuitBuilder usage."""
-        builder = KernelCircuitBuilder()
-        feature_map = builder.input_size(2).n_modes(4).build_feature_map()
-
-        assert feature_map.input_size == 2
-        assert feature_map.circuit.m == 4
-
-    def test_builder_with_device_and_dtype(self):
-        """Test builder with device and dtype configuration."""
-        device = torch.device("cpu")
-        builder = KernelCircuitBuilder()
-        feature_map = (
-            builder
-            .input_size(2)
-            .n_modes(4)
-            .device(device)
-            .dtype(torch.float64)
-            .build_feature_map()
-        )
-
-        assert feature_map.input_size == 2
-        assert feature_map.device == device
-
-    def test_builder_trainable_toggle(self):
-        """Builder can enable or disable training dynamically."""
-        builder = KernelCircuitBuilder()
-        feature_map = (
-            builder.input_size(2).n_modes(4).trainable(False).build_feature_map()
-        )
-
-        assert feature_map.input_size == 2
-        assert not feature_map.is_trainable
-
-        feature_map = (
-            builder
-            .input_size(2)
-            .n_modes(4)
-            .trainable(True, prefix="phi_")
-            .build_feature_map()
-        )
-
-        assert feature_map.is_trainable
-        assert "phi_" in feature_map.trainable_parameters
-
-    def test_builder_build_fidelity_kernel(self):
-        """Test building a FidelityKernel directly."""
-        builder = KernelCircuitBuilder()
-        kernel = builder.input_size(2).n_modes(4).build_fidelity_kernel()
-
-        assert kernel.input_size == 2
-        assert kernel.feature_map.circuit.m == 4
-        assert len(kernel.input_state) == 4
-        assert sum(kernel.input_state) == 2
-        assert kernel.input_state == [1, 0, 1, 0]
-
-    def test_builder_fidelity_kernel_with_custom_input_state(self):
-        """Test building FidelityKernel with custom input state."""
-        builder = KernelCircuitBuilder()
-        custom_state = [2, 0, 0, 0]
-        kernel = (
-            builder
-            .input_size(2)
-            .n_modes(4)
-            .build_fidelity_kernel(input_state=custom_state)
-        )
-
-        assert kernel.input_state == custom_state
-
-    def test_builder_fidelity_kernel_with_shots(self):
-        """Test building FidelityKernel with sampling configuration."""
-        builder = KernelCircuitBuilder()
-        kernel = (
-            builder
-            .input_size(2)
-            .n_modes(4)
-            .build_fidelity_kernel(
-                shots=1000,
-                sampling_method="multinomial",
-                computation_space=ComputationSpace.UNBUNCHED,
-            )
-        )
-
-        assert kernel.shots == 1000
-        assert kernel.sampling_method == "multinomial"
-        assert kernel.computation_space is ComputationSpace.UNBUNCHED
-
-    def test_builder_default_values(self):
-        """Test builder with default values for optional parameters."""
-        builder = KernelCircuitBuilder()
-        feature_map = builder.input_size(2).build_feature_map()
-
-        assert feature_map.input_size == 2
-        # Should use defaults: n_modes = max(input_size + 1, 4) = 4
-        assert feature_map.circuit.m == 4
-
-    def test_builder_angle_encoding_configuration(self):
-        builder = KernelCircuitBuilder()
-        feature_map = (
-            builder
-            .input_size(3)
-            .n_modes(4)
-            .angle_encoding(scale=0.5)
-            .build_feature_map()
-        )
-
-        x = torch.tensor([0.1, 0.2, 0.3], dtype=torch.float32)
-        encoded = feature_map._encode_x(x)
-
-        assert encoded.shape == (3,)
-
-        expected = torch.tensor([0.05, 0.1, 0.15], dtype=torch.float32)
-        assert torch.allclose(encoded.detach(), expected, atol=1e-6)
+class TestKernelConstructionConsistency:
+    """Test integration using the supported circuit construction APIs."""
 
     def test_kernel_supports_entangling_layer(self):
         builder = CircuitBuilder(n_modes=4)
@@ -1108,32 +819,6 @@ class TestKernelCircuitBuilder:
 
         assert K.shape == (3, 3)
         assert torch.isfinite(K).all()
-
-    def test_builder_missing_input_size(self):
-        """Test builder error when input_size is not specified."""
-        builder = KernelCircuitBuilder()
-
-        with pytest.raises(ValueError, match="Input size must be specified"):
-            builder.n_modes(4).build_feature_map()
-
-    def test_deprecation_warning_on_init(self):
-        """KernelCircuitBuilder emits DeprecationWarning on instantiation."""
-        with pytest.warns(DeprecationWarning, match="KernelCircuitBuilder is deprecated"):
-            KernelCircuitBuilder()
-
-    def test_deprecation_warning_build_feature_map(self):
-        """build_feature_map emits DeprecationWarning."""
-        with pytest.warns(DeprecationWarning, match="KernelCircuitBuilder"):
-            KernelCircuitBuilder().input_size(2).build_feature_map()
-
-    def test_deprecation_warning_build_fidelity_kernel(self):
-        """build_fidelity_kernel emits DeprecationWarning."""
-        with pytest.warns(DeprecationWarning, match="KernelCircuitBuilder"):
-            KernelCircuitBuilder().input_size(2).n_modes(4).build_fidelity_kernel()
-
-
-class TestKernelConstructionConsistency:
-    """Test integration using the supported circuit construction APIs."""
 
     def test_feature_map_unitary_consistency(self):
         """Feature maps built via different APIs share the same topology."""
